Kernel/IntelGraphics: Move auxiliary connectors to a separate directory
