+++ Kernel/Graphics/Intel/Auxiliary/DisplayPortAuxiliaryConnector.h
