+++ Kernel/Graphics/Intel/Auxiliary/GMBusConnector.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <Kernel/Arch/Delay.h>
-#include <Kernel/Graphics/Intel/GMBusConnector.h>
+#include <Kernel/Graphics/Intel/Auxiliary/GMBusConnector.h>
 #include <Kernel/PhysicalAddress.h>
 
 namespace Kernel {
