+++ Kernel/Graphics/Intel/Auxiliary/GMBusConnector.h
