+++ Kernel/Graphics/Intel/NativeDisplayConnector.h
@@ -10,7 +10,6 @@
 #include <Kernel/Graphics/Console/GenericFramebufferConsole.h>
 #include <Kernel/Graphics/Definitions.h>
 #include <Kernel/Graphics/DisplayConnector.h>
-#include <Kernel/Graphics/Intel/GMBusConnector.h>
 #include <Kernel/Library/LockRefPtr.h>
 #include <Kernel/Memory/TypedMapping.h>
